p44pagekite: plan44 customized version of pagekite tool (from libpagekite package)
In addition to standard pagekite tool functionality, the p44pagekite tool can fetch the actual kite operating parameters from a list of config files (-f option), and adds a mechanism to let kites expire at a specified time to implement temporary support connections.
The service needs runit (part of busybox or separate) installed to run as-is.
The run script /etc/service/pagekite/run is configured to use the plan44 pagekite frontent, which is not publicly accessible.
So this package is primarily intended for plan44 related projects. If you want to use it you must modifiy the startup script to use your own pagekitefrontend server.